refactor(metrics): updateMetricsName を独立させる
